makehaikufloppy: delete.
The last change to it was 2009, in which Stephan said "I have no
idea if this even still works." With the advent of packagefs,
it's even more unlikely that it does, and there's no real need for
it anyway.

If someone happes to have a usecase as well as a patch to make it work,
we can consider re-adding it.