fat: Greatly simplify and clean up dosfs_get_file_map().
There's no real need to distinguish between first/middle/last parts of
the file and much of the initialization can be done simpler.

This also checks for overflows when truncating the requested length. On
64 bit platforms this would always happen due to the kernel file map
code requesting (size_t)-1 (i.e. unlimited) extents. This lead to the
file end being reached when building the map from a position unequal 0,
which would happen for files fragmented enough to need more than the
default of 8 supplied vectors. An IO error was returned in that case,
rendering the file partially unreadable.