fat: Fix corruption of file data with removed directory entries.
Directory entries are cached using the block cache, so the cached
blocks need to be discarded when directories are removed/shrunk.
Otherwise freed directory blocks that were reused for file data would
later be clobbered when the block cache was flushed.

The forced cache sync for removable media hid the problem for these
devices. It was easily triggered on non-removable media however.