Move ports and protocols files back to data directory in Haiku package.
These files are required for netresolv functionality, and there is no
real use in modifying them as settings files. Restore the previous
behavior, the files are stored in data and part of the Haiku package.

This means there is no need for a fresh install from image to get the
files anymore.

Fixes #12156.