Fix mess in Tracker add-ons created in hrev48858
Revert the Tracker portions of hrev48858, they were not needed and were not
right either... sorry guys

OpenTargetFolder add-on modified slightly to use the passed in symlink(s) only,
the parent directory ref is unused, it appears to be unneeded here also.

With this and the reverted commits Zip-O-Matic and other add-ons should
work again.

Fixes #12105