Revert 64-bit fixes for input_server.
This patch reverts http://cgit.haiku-os.org/haiku/diff/?id=57ab0395ad31761e27ef6d5aa3af68cc3e4d71b2
It may fix #661. Thanks to kcg369 for pointing it out.

Signed-off-by: Adrien Destugues <pulkomandy@pulkomandy.tk>

The initial issue was the storage of a pointer in an int32 for
input method cookies. This was fixed in two ways:

- The int32 was initally replaced with a void*, which allowed to store a
  complete pointer, but changed the format of the BMessages used for
  communication, breaking Canna.
- Then, the code was changed to not use a pointer and instead use an
  int32 for the cookie (see #8831). However, the int32 was still cast
  into a void* when putting it into a BMessage, so the problem was still
  there for Canna.

This commit reverts the remaining parts of the initial solution and
restores the ABI to use int32 everywhere.

Fixes #661.