Add gcc5 packages for x86_64.
* Beware that installing these packages will replace gcc and syslibs. gcc5 packages
can be used to check that packages still build correctly with gcc5, simply being
places in haikuporter packages folder.
* libstdc++ is configured for "gcc4 compatibility", this means we could switch to
this version, while maintaining binary compatibility. However a switch to the new abi
will eventually require a full rebuild of packages depending on libstdc++. For this
to happen, the affected packages should be buildable with gcc5, which can easily be done
with the addition of these packages. Mainstream packages with up to date versions are
probably already buildable with gcc5, some aren't (ie boost). All haiku-* packages are
to be checked. I'll open a provisioning ticket to track the related activities.