Tracker: Fix column title colors following Set*UIColor merge.
Repair regression where low color for text drawing was not set properly.
Do not use tinted color for default color state.

Prior to this, the hinting font would not respond to being pressed and the
column title background was the wrong color initially, but correct after
resizing (but then wrong again after a redraw on exposure).

In addition, the initial tint values used for the view color were unused,
so I removed them.

Signed-off-by: Augustin Cavalier <waddlesplash@gmail.com>