WebPositive: Fix icon background.
Due to a minor change in the handling of default colors with layouts, the
colors were adopting the system/parent colors due to no colors being
specified by the class.

The icon now adopts the document background color and synchronizes with the
url input background.

Signed-off-by: Augustin Cavalier <waddlesplash@gmail.com>
Fixes #12672. I've had this patch in my inbox for a long time, and totally
forgot about committing it...