MediaPlayer PositionToolTip Color
Previously the layout would crush the default colors of BStringView preventing
BStringView from calling AdoptParentColors() on its own, so we must call it
manually.

In addition, the default tooltip view should fully adopt tooltip colors so
that any colors will default to the desired foreground color (which is the
same as the tooltip text color).

Fixes #12573.

Signed-off-by: Jessica Hamilton <jessica.l.hamilton@gmail.com>