Updated packages to Clipdinger 0.5.1
Additionally to more robust pasting to sprunge (now checks if the
service is operational) the required curl version is now the one
that's actually in the HaikuPorts repo, 7.45.0.
Clipdinger 0.5 depended on a newer version, which rendered it uninstallable
with only the default HaikuPorts repo.

There's a messaging issue that results in pasting to sprunge twice, which
I'm still investigating. Releasing now, as it's not a debilitating bug.